backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-channels==4.0.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
anthropic==0.18.1
pinecone-client==3.0.2
boto3==1.34.0
langchain==0.1.0
langchain-anthropic==0.1.0
langchain-aws==0.1.0
pandas==2.1.4
numpy==1.24.3
requests==2.31.0 